bunch of changes

+++ Modules/ApiPatient/Routes/web.php
@@ -17,18 +17,17 @@ Route::group(['prefix' => 'api'], function () {
                 Route::prefix('problems')->group(
                     function () {
                         Route::get('ccd', 'CcdProblemController@show')->middleware('permission:patientProblem.read');
-                        Route::post(
-                            'ccd',
-                            'CcdProblemController@store'
-                        )->middleware('permission:patientProblem.create');
-                        Route::put(
-                            'ccd/{ccdProblemId}',
-                            'CcdProblemController@update'
-                        )->middleware('permission:patientProblem.update');
-                        Route::delete(
-                            'ccd/{ccdProblemId}',
-                            'CcdProblemController@destroy'
-                        )->middleware('permission:patientProblem.delete');
+                        Route::post('ccd', 'CcdProblemController@store')->middleware('permission:patientProblem.create');
+                        Route::put('ccd/{ccdProblemId}', 'CcdProblemController@update')->middleware('permission:patientProblem.update');
+                        Route::delete('ccd/{ccdProblemId}', 'CcdProblemController@destroy')->middleware('permission:patientProblem.delete');
+                    }
+                );
+
+                Route::prefix('allergies')->group(
+                    function () {
+                        Route::get('', 'AllergyController@index')->middleware('permission:allergy.read');
+                        Route::post('', 'AllergyController@store')->middleware('permission:allergy.create');
+                        Route::delete('{allergyId}', 'AllergyController@destroy')->middleware('permission:allergy.delete');
                     }
                 );
             }
